fileoperations/requirement.txt
blinker==1.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pycparser==2.21
python-dotenv==1.0.0
Werkzeug==2.3.4
